move boilerplate code to own class
